+++ changelog.go
@@ -11,6 +11,9 @@ import (
 	"github.com/spf13/afero"
 )
 
+// Filesystem is an interface of a filesystem.
+//
+// Possible options: [afero.NewOsFs(), afero.NewMemMapFs()].
 type Filesystem interface {
 	Stat(string) (fs.FileInfo, error)
 	Open(string) (afero.File, error)
@@ -25,7 +28,7 @@ type Changelog struct {
 	Releases    Releases
 }
 
-// ToString returns a Markdown formatted Changelog struct
+// ToString returns a Markdown formatted Changelog struct.
 func (c *Changelog) ToString() string {
 	var o []string
 	var defs []string
@@ -59,7 +62,8 @@ func (c *Changelog) ToString() string {
 
 // SaveToFile formats the changelog struct according to a predefined format
 // and prints it to file.
-// Possible options for `filesystems` are: [`afero.NewOsFs()`,`afero.NewMemMapFs()`]
+//
+// Possible options for Filesystem are: [afero.NewOsFs(), afero.NewMemMapFs()].
 func (c *Changelog) SaveToFile(filesystem Filesystem, filepath string) error {
 	f, err := filesystem.Create(filepath)
 	if err != nil {
@@ -79,7 +83,7 @@ func (c *Changelog) SaveToFile(filesystem Filesystem, filepath string) error {
 	return nil
 }
 
-// NewChangelog returns an empty changelog
+// NewChangelog returns an empty changelog.
 func NewChangelog() *Changelog {
 	c := new(Changelog)
 	c.Releases = make(Releases, 0)
@@ -87,32 +91,37 @@ func NewChangelog() *Changelog {
 	return c
 }
 
+// SetTitle updates a title of the changelog.
 func (c *Changelog) SetTitle(title string) {
 	*c.Title = title
 }
 
+// SetDescription updates a description of the changelog.
 func (c *Changelog) SetDescription(description string) {
 	*c.Description = description
 }
 
+// SetUnreleasedURL configures a markdown URL for an Unreleased section.
 func (c *Changelog) SetUnreleasedURL(link string) error {
-	u, err := url.Parse(link)
+	_, err := url.Parse(link)
 	if err != nil {
 		return err
 	}
-	urlString := u.String()
 
 	if c.Unreleased == nil {
 		c.Unreleased = &Release{
-			URL: &urlString,
+			URL: &link,
 		}
 	} else {
-		*c.Unreleased.URL = link
+		c.Unreleased.URL = &link
 	}
 
 	return nil
 }
 
+// AddUnreleasedChange adds a scoped change to Unreleased section.
+//
+// Supported scopes: [added, changed, deprecated, removed, fixed, security].
 func (c *Changelog) AddUnreleasedChange(scope string, change string) error {
 	if c.Unreleased == nil {
 		c.Unreleased = &Release{
@@ -127,12 +136,55 @@ func (c *Changelog) AddUnreleasedChange(scope string, change string) error {
 	return c.Unreleased.Changes.AddChange(scope, change)
 }
 
+// GetRelease returns a release for a provided version.
+//
+// This is a helper function that wraps Releases.GetRelease function.
 func (c *Changelog) GetRelease(version string) *Release {
-	for _, r := range c.Releases {
-		if *r.Version == version {
-			return r
-		}
+	return c.Releases.GetRelease(version)
+}
+
+// CreateReleaseFromUnreleased creates a new release with all the changes from Unreleased section.
+// This will also cleanup the Unreleased section.
+func (c *Changelog) CreateReleaseFromUnreleased(version, date string) (*Release, error) {
+	r, err := c.CreateRelease(version, date)
+	if err != nil {
+		return nil, err
 	}
 
-	return nil
+	r.Changes = c.Unreleased.Changes
+	c.Unreleased.Changes = nil
+
+	return r, nil
+}
+
+// CreateReleaseFromUnreleased creates a new release with all the changes from Unreleased section.
+// This will also cleanup the Unreleased section.
+// Identical to CreateReleaseFromUnreleased but with an extra step of adding a URL to the release.
+func (c *Changelog) CreateReleaseFromUnreleasedWithURL(version, date, url string) (*Release, error) {
+	r, err := c.CreateReleaseFromUnreleased(version, date)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := r.SetURL(url); err != nil {
+		return nil, err
+	}
+
+	return r, nil
+}
+
+// CreateRelease creates new empty release.
+//
+// This is a helper function that wraps Releases.CreateRelease function.
+func (c *Changelog) CreateRelease(version, date string) (*Release, error) {
+	return c.Releases.CreateRelease(version, date)
+}
+
+// CreateReleaseWithURL creates new empty release.
+//
+// This is a helper function that wraps Releases.CreateReleaseWithURL function.
+//
+// Identical to CreateRelease but with an extra step of adding a URL to the release.
+func (c *Changelog) CreateReleaseWithURL(version, date, url string) (*Release, error) {
+	return c.Releases.CreateReleaseWithURL(version, date, url)
 }